+++ src/index.css
@@ -12,3 +12,5 @@ code {
   font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New','SEGA LOGO FONT',
     monospace;
 }
+
+html, body {margin: 0; height: 100%; overflow: hidden}
